+++ mithril-stm/src/circuits/halo2_ivc/gadget.rs
@@ -10,13 +10,13 @@ use super::{
     Accumulator, ArithInstructions, AssertionInstructions, AssignedAccumulator, AssignedBit,
     AssignedForeignPoint, AssignedNative, AssignedNativePoint, AssignedScalarOfNativeCurve,
     AssignedVk, AssignmentInstructions, BinaryInstructions, CERTIFICATE_VERIFICATION_KEY_NAME,
-    CircuitCurve, ComposableChip, ConstraintSystem, ControlFlowInstructions,
-    ConversionInstructions, EccChip, EccInstructions, EmulatedCurve, EqualityInstructions, Error,
-    EvaluationDomain, ForeignEccChip, HashInstructions, IVC_VERIFICATION_KEY_NAME, IvcNativeGadget,
-    Jubjub, Layouter, NativeChip, NativeField, NativeGadget, P2RDecompositionChip,
-    PREIMAGE_CURRENT_EPOCH_BYTES, PREIMAGE_NEXT_MERKLE_TREE_COMMITMENT_BYTES,
+    CircuitCurve, CircuitCurveTrait, CircuitValue, ComposableChip, ConstraintSystem,
+    ControlFlowInstructions, ConversionInstructions, EccChip, EccInstructions, EmulatedCurve,
+    EqualityInstructions, Error, EvaluationDomain, ForeignEccChip, HashInstructions,
+    IVC_VERIFICATION_KEY_NAME, IvcNativeGadget, Layouter, NativeChip, NativeField, NativeGadget,
+    P2RDecompositionChip, PREIMAGE_CURRENT_EPOCH_BYTES, PREIMAGE_NEXT_MERKLE_TREE_COMMITMENT_BYTES,
     PREIMAGE_NEXT_PROTOCOL_PARAMETERS_BYTES, PoseidonChip, PublicInputInstructions,
-    RECURSIVE_CIRCUIT_DEGREE, RecursiveEmulation, Value, VerifierGadget, ZeroInstructions,
+    RECURSIVE_CIRCUIT_DEGREE, RecursiveEmulation, VerifierGadget, ZeroInstructions,
     config::IvcConfig,
     errors::{IvcCircuitError, to_synthesis_error},
     state::{
@@ -28,7 +28,7 @@ use super::{
 pub struct IvcGadget {
     pub(crate) core_decomp_chip: P2RDecompositionChip<NativeField>,
     pub(crate) native_gadget: IvcNativeGadget,
-    pub(crate) jubjub_chip: EccChip<Jubjub>,
+    pub(crate) jubjub_chip: EccChip<CircuitCurve>,
     pub(crate) poseidon_chip: PoseidonChip<NativeField>,
     pub(crate) sha2_256_chip: Sha256Chip<NativeField>,
     pub(crate) bls12_381_chip:
@@ -47,7 +47,7 @@ impl IvcGadget {
             &(RECURSIVE_CIRCUIT_DEGREE as usize - 1),
         );
         let native_gadget = NativeGadget::new(core_decomp_chip.clone(), native_chip.clone());
-        let jubjub_chip = EccChip::<Jubjub>::new(&config.jubjub_config, &native_gadget);
+        let jubjub_chip = EccChip::<CircuitCurve>::new(&config.jubjub_config, &native_gadget);
         let bls12_381_chip: ForeignEccChip<_, EmulatedCurve, EmulatedCurve, _, _> =
             { ForeignEccChip::new(&config.bls12_381_config, &native_gadget, &native_gadget) };
         let poseidon_chip = PoseidonChip::new(&config.poseidon_config, &native_chip);
@@ -69,7 +69,7 @@ impl IvcGadget {
     pub fn assign_global_as_public_input(
         &self,
         layouter: &mut impl Layouter<NativeField>,
-        global: &Value<Global>,
+        global: &CircuitValue<Global>,
         certificate_circuit_domain_and_constraint_system: &(
             EvaluationDomain<NativeField>,
             ConstraintSystem<NativeField>,
@@ -145,7 +145,7 @@ impl IvcGadget {
     pub fn assign_state(
         &self,
         layouter: &mut impl Layouter<NativeField>,
-        state: &Value<State>,
+        state: &CircuitValue<State>,
     ) -> Result<AssignedState, Error> {
         let values = state
             .clone()
@@ -216,7 +216,7 @@ impl IvcGadget {
     pub fn assign_witness(
         &self,
         layouter: &mut impl Layouter<NativeField>,
-        witness: &Value<Witness>,
+        witness: &CircuitValue<Witness>,
     ) -> Result<AssignedWitness, Error> {
         let genesis_signature = {
             let s: AssignedScalarOfNativeCurve<_> = self.jubjub_chip.assign(
@@ -297,7 +297,7 @@ impl IvcGadget {
             .assign_fixed(layouter, DOMAIN_SEPARATION_TAG_STANDARD_SIGNATURE.0)?;
         let generator: AssignedNativePoint<_> = self.jubjub_chip.assign_fixed(
             layouter,
-            <Jubjub as CircuitCurve>::CryptographicGroup::generator(),
+            <CircuitCurve as CircuitCurveTrait>::CryptographicGroup::generator(),
         )?;
 
         let cap_r = self.jubjub_chip.msm(
@@ -587,9 +587,9 @@ impl IvcGadget {
         is_not_genesis: &AssignedBit<NativeField>,
         state: &AssignedState,
         witness: &AssignedWitness,
-        certificate_proof: &Value<Vec<u8>>,
-        ivc_proof: &Value<Vec<u8>>,
-        acc_value: &Value<Accumulator<RecursiveEmulation>>,
+        certificate_proof: &CircuitValue<Vec<u8>>,
+        ivc_proof: &CircuitValue<Vec<u8>>,
+        acc_value: &CircuitValue<Accumulator<RecursiveEmulation>>,
     ) -> Result<AssignedAccumulator<RecursiveEmulation>, Error> {
         let id_point: AssignedForeignPoint<_, _, _> = self
             .bls12_381_chip